Creacion de todas las tablas

+++ Entrega 2/src/learningPaths/Recurso.java
@@ -13,7 +13,7 @@ public class Recurso extends Actividad {
 	public Actividad prerequisistos;
 	public Actividad sugeridos;
 	public ArrayList<String> resenias;
-	public double rating;
+	public float rating;
 	public ArrayList<Integer> ratings;
 	public float tiempoLimite;
 	public boolean completado;
